requirements.txt
python-dotenv
squareup
fastapi
uvicorn
boto3
